cmd/build: Make symlinks by default
* Parse $VOLTPATH/config.toml
* `volt build` makes symlinks under ~/.vim/pack/volt/opt/ by default
