+++ cmd/build.go
@@ -1,32 +1,18 @@
 package cmd
 
 import (
-	"encoding/json"
 	"errors"
 	"flag"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-	"time"
 
-	"github.com/hashicorp/go-multierror"
-	"github.com/vim-volt/volt/fileutil"
-	"github.com/vim-volt/volt/lockjson"
+	"github.com/vim-volt/volt/cmd/builder"
+	"github.com/vim-volt/volt/cmd/buildinfo"
+	"github.com/vim-volt/volt/config"
 	"github.com/vim-volt/volt/logger"
 	"github.com/vim-volt/volt/pathutil"
-	"github.com/vim-volt/volt/plugconf"
 	"github.com/vim-volt/volt/transaction"
-
-	"gopkg.in/src-d/go-git.v4"
-	"gopkg.in/src-d/go-git.v4/plumbing"
-	"gopkg.in/src-d/go-git.v4/plumbing/object"
 )
-
-var BuildModeInvalidType = os.ModeSymlink | os.ModeNamedPipe | os.ModeSocket | os.ModeDevice
 
 func init() {
 	cmdMap["build"] = &buildCmd{}
@@ -94,144 +80,45 @@ func (cmd *buildCmd) Run(args []string) int {
 	return 0
 }
 
-const currentBuildInfoVersion = 1
-
-type buildInfoType struct {
-	Repos   biReposList `json:"repos"`
-	Version int64       `json:"version"`
-}
-
-type biReposList []biRepos
-
-type biRepos struct {
-	Type          lockjson.ReposType `json:"type"`
-	Path          string             `json:"path"`
-	Version       string             `json:"version"`
-	Files         biFileMap          `json:"files,omitempty"`
-	DirtyWorktree bool               `json:"dirty_worktree,omitempty"`
-}
-
-// key: filepath, value: version
-type biFileMap map[string]string
-
-func (cmd *buildCmd) readBuildInfo() (*buildInfoType, error) {
-	// Return initial build-info.json struct
-	// if the file does not exist
-	file := pathutil.BuildInfoJSON()
-	if !pathutil.Exists(file) {
-		return &buildInfoType{}, nil
-	}
-
-	// Read build-info.json
-	bytes, err := ioutil.ReadFile(file)
-	if err != nil {
-		return nil, err
-	}
-	var buildInfo buildInfoType
-	err = json.Unmarshal(bytes, &buildInfo)
-	if err != nil {
-		return nil, err
-	}
-
-	// Validate build-info.json
-	err = buildInfo.validate()
-	if err != nil {
-		return nil, errors.New("validation failed: build-info.json: " + err.Error())
-	}
-
-	return &buildInfo, nil
-}
-
-func (buildInfo *buildInfoType) write() error {
-	// Validate build-info.json
-	err := buildInfo.validate()
-	if err != nil {
-		return errors.New("validation failed: build-info.json: " + err.Error())
-	}
-
-	// Write to build-info.json
-	bytes, err := json.MarshalIndent(buildInfo, "", "  ")
-	if err != nil {
-		return err
-	}
-	return ioutil.WriteFile(pathutil.BuildInfoJSON(), bytes, 0644)
-}
-
-func (buildInfo *buildInfoType) validate() error {
-	// Validate if repos do not have duplicate repository
-	dupRepos := make(map[string]bool, len(buildInfo.Repos))
-	for i := range buildInfo.Repos {
-		r := &buildInfo.Repos[i]
-		if _, exists := dupRepos[r.Path]; exists {
-			return errors.New("duplicate repos: " + r.Path)
-		}
-		dupRepos[r.Path] = true
-	}
-	return nil
-}
-
-func (reposList *biReposList) findByReposPath(reposPath string) *biRepos {
-	for i := range *reposList {
-		repos := &(*reposList)[i]
-		if repos.Path == reposPath {
-			return repos
-		}
-	}
-	return nil
-}
-
-func (reposList *biReposList) removeByReposPath(reposPath string) {
-	for i := range *reposList {
-		repos := &(*reposList)[i]
-		if repos.Path == reposPath {
-			*reposList = append((*reposList)[:i], (*reposList)[i+1:]...)
-			break
-		}
-	}
-}
+const currentBuildInfoVersion = 2
 
 func (cmd *buildCmd) doBuild(full bool) error {
-	// Exit if vim executable was not found in PATH
-	if _, err := pathutil.VimExecutable(); err != nil {
-		return err
-	}
-
-	vimDir := pathutil.VimDir()
-	optDir := pathutil.VimVoltOptDir()
-
-	// Read lock.json
-	lockJSON, err := lockjson.Read()
+	// Read config.toml
+	cfg, err := config.Read()
 	if err != nil {
-		return errors.New("could not read lock.json: " + err.Error())
+		return errors.New("could not read config.toml: " + err.Error())
 	}
 
-	// Get current profile's repos list
-	profile, reposList, err := cmd.getCurrentProfileAndReposList(lockJSON)
+	// Get builder
+	builder, err := builder.Get(cfg.Build.Strategy)
 	if err != nil {
 		return err
 	}
 
 	// Read ~/.vim/pack/volt/opt/build-info.json
-	buildInfo, err := cmd.readBuildInfo()
+	buildInfo, err := buildinfo.Read()
 	if err != nil {
 		return err
 	}
 
-	// Do full build when build-info.json's version is different
-	if buildInfo.Version != currentBuildInfoVersion {
+	// Do full build when build-info.json's version or strategy is different
+	if buildInfo.Version != currentBuildInfoVersion ||
+		buildInfo.Strategy != cfg.Build.Strategy {
 		full = true
 	}
 	buildInfo.Version = currentBuildInfoVersion
+	buildInfo.Strategy = cfg.Build.Strategy
 
 	// Put repos into map to be able to search with O(1).
 	// Use empty build-info.json map if the -full option was given
 	// because the repos info is unnecessary because it is not referenced.
-	var buildReposMap map[string]*biRepos
+	var buildReposMap map[string]*buildinfo.Repos
+	optDir := pathutil.VimVoltOptDir()
 	if full {
-		buildReposMap = make(map[string]*biRepos)
+		buildReposMap = make(map[string]*buildinfo.Repos)
 		logger.Info("Full building " + optDir + " directory ...")
 	} else {
-		buildReposMap = make(map[string]*biRepos, len(buildInfo.Repos))
+		buildReposMap = make(map[string]*buildinfo.Repos, len(buildInfo.Repos))
 		for i := range buildInfo.Repos {
 			repos := &buildInfo.Repos[i]
 			buildReposMap[repos.Path] = repos
@@ -242,712 +129,11 @@ func (cmd *buildCmd) doBuild(full bool) error {
 	// Remove ~/.vim/pack/volt/ if -full option was given
 	if full {
 		vimVoltDir := pathutil.VimVoltDir()
-		err = os.RemoveAll(vimVoltDir)
-		if err != nil {
-			return errors.New("failed to remove " + vimVoltDir + ": " + err.Error())
+		os.RemoveAll(vimVoltDir)
+		if pathutil.Exists(vimVoltDir) {
+			return errors.New("failed to remove " + vimVoltDir)
 		}
 	}
 
-	logger.Info("Installing vimrc and gvimrc ...")
-
-	vimrcPath := filepath.Join(vimDir, pathutil.Vimrc)
-	gvimrcPath := filepath.Join(vimDir, pathutil.Gvimrc)
-	err = cmd.installVimrcAndGvimrc(
-		lockJSON.CurrentProfileName, vimrcPath, gvimrcPath, profile.UseVimrc, profile.UseGvimrc,
-	)
-	if err != nil {
-		return err
-	}
-
-	// Mkdir opt dir
-	os.MkdirAll(optDir, 0755)
-	if !pathutil.Exists(optDir) {
-		return errors.New("could not create " + optDir)
-	}
-
-	reposDirList, err := ioutil.ReadDir(pathutil.VimVoltOptDir())
-	if err != nil {
-		return err
-	}
-
-	// Copy volt repos files to optDir
-	copyDone, copyCount := cmd.copyReposList(buildReposMap, reposList, optDir)
-
-	// Remove vim repos not found in lock.json current repos list
-	removeDone, removeCount := cmd.removeReposList(reposList, reposDirList)
-
-	// Wait copy
-	var copyModified bool
-	copyErr := cmd.waitCopyRepos(copyDone, copyCount, func(result *actionReposResult) error {
-		logger.Info("Installing " + string(result.repos.Type) + " repository " + result.repos.Path + " ... Done.")
-		// Construct buildInfo from the result
-		cmd.constructBuildInfo(buildInfo, result)
-		copyModified = true
-		return nil
-	})
-
-	// Wait remove
-	var removeModified bool
-	removeErr := cmd.waitRemoveRepos(removeDone, removeCount, func(result *actionReposResult) {
-		// Remove the repository from buildInfo
-		buildInfo.Repos.removeByReposPath(result.repos.Path)
-		removeModified = true
-	})
-
-	// Handle copy & remove errors
-	if copyErr != nil || removeErr != nil {
-		return multierror.Append(copyErr, removeErr).ErrorOrNil()
-	}
-
-	// Write bundled plugconf file
-	content, merr := plugconf.GenerateBundlePlugconf(reposList)
-	if merr.ErrorOrNil() != nil {
-		// Return vim script parse errors
-		return merr
-	}
-	os.MkdirAll(filepath.Dir(pathutil.BundledPlugConf()), 0755)
-	err = ioutil.WriteFile(pathutil.BundledPlugConf(), content, 0644)
-	if err != nil {
-		return err
-	}
-
-	// Write to build-info.json if buildInfo was modified
-	if copyModified || removeModified {
-		err = buildInfo.write()
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (cmd *buildCmd) installVimrcAndGvimrc(profileName, vimrcPath, gvimrcPath string, useVimrc, useGvimrc bool) error {
-	// Save old vimrc file as {vimrc}.bak
-	vimrcInfo, err := os.Stat(vimrcPath)
-	if err != nil && !os.IsNotExist(err) {
-		return err
-	}
-	vimrcExists := !os.IsNotExist(err)
-	if vimrcExists {
-		err = fileutil.CopyFile(vimrcPath, vimrcPath+".bak", make([]byte, vimrcInfo.Size()), vimrcInfo.Mode())
-		if err != nil {
-			return err
-		}
-	}
-	defer os.Remove(vimrcPath + ".bak")
-
-	// Install vimrc
-	err = cmd.installRCFile(
-		profileName,
-		pathutil.ProfileVimrc,
-		vimrcPath,
-		useVimrc,
-	)
-	if err != nil {
-		return err
-	}
-
-	// Install gvimrc
-	err = cmd.installRCFile(
-		profileName,
-		pathutil.ProfileGvimrc,
-		gvimrcPath,
-		useGvimrc,
-	)
-	if err != nil {
-		// Restore old vimrc
-		if vimrcExists {
-			err2 := os.Rename(vimrcPath+".bak", vimrcPath)
-			if err2 != nil {
-				return multierror.Append(err, err2)
-			}
-		} else {
-			err2 := os.Remove(vimrcPath)
-			if err2 != nil {
-				return multierror.Append(err, err2)
-			}
-		}
-		return err
-	}
-	return nil
-}
-
-func (cmd *buildCmd) installRCFile(profileName, srcRCFileName, dst string, install bool) error {
-	src := filepath.Join(pathutil.RCDir(profileName), srcRCFileName)
-
-	// Return error if destination file does not have magic comment
-	if pathutil.Exists(dst) {
-		// If the file does not have magic comment
-		if !cmd.hasMagicComment(dst) {
-			if !pathutil.Exists(src) {
-				return nil
-			}
-			return errors.New("'" + dst + "' does not have magic comment")
-		}
-	}
-
-	// Remove destination (~/.vim/vimrc or ~/.vim/gvimrc)
-	os.Remove(dst)
-	if pathutil.Exists(dst) {
-		return errors.New("failed to remove " + dst)
-	}
-
-	// Skip if use_vimrc/use_gvimrc is false or rc file does not exist
-	if !install || !pathutil.Exists(src) {
-		return nil
-	}
-
-	return cmd.copyFileWithMagicComment(src, dst)
-}
-
-const magicComment = "\" NOTE: this file was generated by volt. please modify original file.\n"
-const magicCommentNext = "\" Original file: %s\n\n"
-
-// Return error if the magic comment does not exist
-func (*buildCmd) hasMagicComment(dst string) bool {
-	r, err := os.Open(dst)
-	if err != nil {
-		return false
-	}
-	defer r.Close()
-
-	magic := []byte(magicComment)
-	read := make([]byte, len(magic))
-	n, err := r.Read(read)
-	if err != nil || n < len(magicComment) {
-		return false
-	}
-
-	for i := range magic {
-		if magic[i] != read[i] {
-			return false
-		}
-	}
-	return true
-}
-
-func (*buildCmd) copyFileWithMagicComment(src, dst string) (err error) {
-	r, err := os.Open(src)
-	if err != nil {
-		return
-	}
-	defer func() {
-		if e := r.Close(); e != nil {
-			err = e
-		}
-	}()
-
-	os.MkdirAll(filepath.Dir(dst), 0755)
-	w, err := os.Create(dst)
-	if err != nil {
-		return
-	}
-	defer func() {
-		if e := w.Close(); e != nil {
-			err = e
-		}
-	}()
-
-	_, err = w.Write([]byte(magicComment))
-	if err != nil {
-		return
-	}
-	_, err = w.Write([]byte(fmt.Sprintf(magicCommentNext, src)))
-	if err != nil {
-		return
-	}
-
-	_, err = io.Copy(w, r)
-	return
-}
-
-type actionReposResult struct {
-	err   error
-	repos *lockjson.Repos
-	files biFileMap
-}
-
-func (cmd *buildCmd) copyReposList(buildReposMap map[string]*biRepos, reposList []lockjson.Repos, optDir string) (chan actionReposResult, int) {
-	copyDone := make(chan actionReposResult, len(reposList))
-	copyCount := 0
-	for i := range reposList {
-		if reposList[i].Type == lockjson.ReposGitType {
-			n, err := cmd.copyReposGit(&reposList[i], buildReposMap[reposList[i].Path], copyDone)
-			if err != nil {
-				copyDone <- actionReposResult{
-					err:   errors.New("failed to copy " + string(reposList[i].Type) + " repos: " + err.Error()),
-					repos: &reposList[i],
-				}
-			}
-			copyCount += n
-		} else if reposList[i].Type == lockjson.ReposStaticType {
-			copyCount += cmd.copyReposStatic(&reposList[i], buildReposMap[reposList[i].Path], optDir, copyDone)
-		} else {
-			copyDone <- actionReposResult{
-				err:   errors.New("invalid repository type: " + string(reposList[i].Type)),
-				repos: &reposList[i],
-			}
-		}
-	}
-	return copyDone, copyCount
-}
-
-func (cmd *buildCmd) copyReposGit(repos *lockjson.Repos, buildRepos *biRepos, done chan actionReposResult) (int, error) {
-	// Open ~/volt/repos/{repos}
-	src := pathutil.FullReposPathOf(repos.Path)
-	r, err := git.PlainOpen(src)
-	if err != nil {
-		return 0, errors.New("failed to open repository: " + err.Error())
-	}
-
-	cfg, err := r.Config()
-	if err != nil {
-		return 0, errors.New("failed to get repository config: " + err.Error())
-	}
-
-	isClean := false
-	if wt, err := r.Worktree(); err == nil {
-		if st, err := wt.Status(); err == nil && st.IsClean() {
-			isClean = true
-		}
-	}
-
-	if cmd.hasChangedGitRepos(repos, buildRepos, !isClean) {
-		// Copy files from .git/objects/... when:
-		// * bare repository
-		// * or worktree is clean
-		copyFromGitObjects := cfg.Core.IsBare || isClean
-		go cmd.updateGitRepos(repos, r, copyFromGitObjects, done)
-		return 1, nil
-	}
-	return 0, nil
-}
-
-func (cmd *buildCmd) copyReposStatic(repos *lockjson.Repos, buildRepos *biRepos, optDir string, done chan actionReposResult) int {
-	if cmd.hasChangedStaticRepos(repos, buildRepos, optDir) {
-		go cmd.updateStaticRepos(repos, done)
-		return 1
-	}
-	return 0
-}
-
-// Remove vim repos not found in lock.json current repos list
-func (cmd *buildCmd) removeReposList(reposList lockjson.ReposList, reposDirList []os.FileInfo) (chan actionReposResult, int) {
-	removeList := make([]string, 0, len(reposList))
-	for i := range reposDirList {
-		reposPath := pathutil.UnpackPathOf(reposDirList[i].Name())
-		if !reposList.Contains(reposPath) {
-			removeList = append(removeList, reposPath)
-		}
-	}
-	removeDone := make(chan actionReposResult, len(removeList))
-	for i := range removeList {
-		go func(reposPath string) {
-			err := os.RemoveAll(pathutil.PackReposPathOf(reposPath))
-			logger.Info("Removing " + reposPath + " ... Done.")
-			removeDone <- actionReposResult{
-				err:   err,
-				repos: &lockjson.Repos{Path: reposPath},
-			}
-		}(removeList[i])
-	}
-	return removeDone, len(removeList)
-}
-
-func (*buildCmd) waitCopyRepos(copyDone chan actionReposResult, copyCount int, callback func(*actionReposResult) error) *multierror.Error {
-	var merr *multierror.Error
-	for i := 0; i < copyCount; i++ {
-		result := <-copyDone
-		if result.err != nil {
-			merr = multierror.Append(
-				merr,
-				errors.New(
-					"failed to copy repository '"+result.repos.Path+
-						"': "+result.err.Error()))
-		} else {
-			err := callback(&result)
-			if err != nil {
-				merr = multierror.Append(merr, err)
-			}
-		}
-	}
-	return merr
-}
-
-func (*buildCmd) constructBuildInfo(buildInfo *buildInfoType, result *actionReposResult) {
-	if result.repos.Type == lockjson.ReposGitType {
-		r := buildInfo.Repos.findByReposPath(result.repos.Path)
-		if r != nil {
-			r.Version = result.repos.Version
-			r.Files = result.files
-		} else {
-			buildInfo.Repos = append(
-				buildInfo.Repos,
-				biRepos{
-					Type:    lockjson.ReposGitType,
-					Path:    result.repos.Path,
-					Version: result.repos.Version,
-					Files:   result.files,
-				},
-			)
-		}
-	} else if result.repos.Type == lockjson.ReposStaticType {
-		r := buildInfo.Repos.findByReposPath(result.repos.Path)
-		if r != nil {
-			r.Version = time.Now().Format(time.RFC3339)
-			r.Files = result.files
-		} else {
-			buildInfo.Repos = append(
-				buildInfo.Repos,
-				biRepos{
-					Type:    lockjson.ReposStaticType,
-					Path:    result.repos.Path,
-					Version: time.Now().Format(time.RFC3339),
-					Files:   result.files,
-				},
-			)
-		}
-	} else {
-		logger.Error("Unknown repos type (" + string(result.repos.Type) + ")")
-	}
-}
-
-func (*buildCmd) waitRemoveRepos(removeDone chan actionReposResult, removeCount int, callback func(result *actionReposResult)) *multierror.Error {
-	var merr *multierror.Error
-	for i := 0; i < removeCount; i++ {
-		result := <-removeDone
-		if result.err != nil {
-			target := "files"
-			if result.repos != nil {
-				target = result.repos.Path
-			}
-			merr = multierror.Append(
-				merr, errors.New(
-					"Failed to remove "+target+
-						": "+result.err.Error()))
-		} else {
-			callback(&result)
-		}
-	}
-	return merr
-}
-
-func (*buildCmd) getCurrentProfileAndReposList(lockJSON *lockjson.LockJSON) (*lockjson.Profile, lockjson.ReposList, error) {
-	// Find current profile
-	profile, err := lockJSON.Profiles.FindByName(lockJSON.CurrentProfileName)
-	if err != nil {
-		// this must not be occurred because lockjson.Read()
-		// validates that the matching profile exists
-		return nil, nil, err
-	}
-
-	reposList, err := lockJSON.GetReposListByProfile(profile)
-	return profile, reposList, err
-}
-
-func (*buildCmd) getLatestModTime(path string) (time.Time, error) {
-	mtime := time.Unix(0, 0)
-	err := filepath.Walk(path, func(_ string, fi os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-		t := fi.ModTime()
-		if mtime.Before(t) {
-			mtime = t
-		}
-		return nil
-	})
-	if err != nil {
-		return time.Now(), errors.New("failed to readdir: " + err.Error())
-	}
-	return mtime, nil
-}
-
-func (*buildCmd) hasChangedGitRepos(repos *lockjson.Repos, buildRepos *biRepos, isDirty bool) bool {
-	if buildRepos == nil { // Full build
-		return true
-	}
-	if repos.Version != buildRepos.Version {
-		return true
-	}
-	if buildRepos.DirtyWorktree || isDirty {
-		return true
-	}
-	return false
-}
-
-// Remove ~/.vim/volt/opt/{repos} and copy from ~/volt/repos/{repos}
-func (cmd *buildCmd) updateGitRepos(repos *lockjson.Repos, r *git.Repository, copyFromGitObjects bool, done chan actionReposResult) {
-	src := pathutil.FullReposPathOf(repos.Path)
-	dst := pathutil.PackReposPathOf(repos.Path)
-
-	// Remove ~/.vim/volt/opt/{repos}
-	// TODO: Do not remove here, copy newer files only after
-	err := os.RemoveAll(dst)
-	if err != nil {
-		done <- actionReposResult{
-			err:   errors.New("failed to remove repository: " + err.Error()),
-			repos: repos,
-		}
-		return
-	}
-
-	if copyFromGitObjects {
-		logger.Debug("Copy from git objects: " + repos.Path)
-		cmd.updateBareGitRepos(r, src, dst, repos, done)
-	} else {
-		logger.Debug("Copy from filesystem: " + repos.Path)
-		cmd.updateNonBareGitRepos(r, src, dst, repos, done)
-	}
-}
-
-func (cmd *buildCmd) updateBareGitRepos(r *git.Repository, src, dst string, repos *lockjson.Repos, done chan actionReposResult) {
-	// Get locked commit hash
-	commit := plumbing.NewHash(repos.Version)
-	commitObj, err := r.CommitObject(commit)
-	if err != nil {
-		done <- actionReposResult{
-			err:   errors.New("failed to get HEAD commit object: " + err.Error()),
-			repos: repos,
-		}
-		return
-	}
-
-	// Get tree hash of commit hash
-	tree, err := r.TreeObject(commitObj.TreeHash)
-	if err != nil {
-		done <- actionReposResult{
-			err:   errors.New("failed to get tree " + commit.String() + ": " + err.Error()),
-			repos: repos,
-		}
-		return
-	}
-
-	// Copy files
-	files := make(biFileMap, 512)
-	err = tree.Files().ForEach(func(file *object.File) error {
-		osMode, err := file.Mode.ToOSFileMode()
-		if err != nil {
-			return errors.New("failed to convert file mode: " + err.Error())
-		}
-
-		contents, err := file.Contents()
-		if err != nil {
-			return errors.New("failed to get file contents: " + err.Error())
-		}
-
-		filename := filepath.Join(dst, file.Name)
-		os.MkdirAll(filepath.Dir(filename), 0755)
-		ioutil.WriteFile(filename, []byte(contents), osMode)
-
-		files[file.Name] = file.Hash.String() // blob hash
-		return nil
-	})
-	if err != nil {
-		done <- actionReposResult{
-			err:   err,
-			repos: repos,
-		}
-		return
-	}
-
-	// Do ":helptags" to generate tags file
-	err = cmd.helptags(repos.Path)
-	if err != nil {
-		done <- actionReposResult{
-			err:   err,
-			repos: repos,
-		}
-		return
-	}
-
-	done <- actionReposResult{
-		err:   nil,
-		repos: repos,
-		files: files,
-	}
-}
-
-func (cmd *buildCmd) updateNonBareGitRepos(r *git.Repository, src, dst string, repos *lockjson.Repos, done chan actionReposResult) {
-	files, err := ioutil.ReadDir(src)
-	if err != nil {
-		done <- actionReposResult{
-			err:   err,
-			repos: repos,
-		}
-		return
-	}
-
-	buf := make([]byte, 32*1024)
-	created := make(map[string]bool, len(files))
-	for _, file := range files {
-		// Skip ".git" and ".gitignore"
-		if file.Name() == ".git" || file.Name() == ".gitignore" {
-			continue
-		}
-		if file.Mode()&BuildModeInvalidType != 0 {
-			// Currenly skip the invalid files...
-			continue
-		}
-		if !created[dst] {
-			os.MkdirAll(dst, 0755)
-			created[dst] = true
-		}
-		from := filepath.Join(src, file.Name())
-		to := filepath.Join(dst, file.Name())
-		var err error
-		if file.IsDir() {
-			err = fileutil.TryLinkDir(from, to, buf, file.Mode(), BuildModeInvalidType)
-		} else {
-			err = fileutil.TryLinkFile(from, to, buf, file.Mode())
-		}
-		if err != nil {
-			done <- actionReposResult{
-				err:   err,
-				repos: repos,
-			}
-			return
-		}
-	}
-
-	err = cmd.helptags(repos.Path)
-	if err != nil {
-		done <- actionReposResult{
-			err:   err,
-			repos: repos,
-		}
-		return
-	}
-
-	done <- actionReposResult{
-		err:   nil,
-		repos: repos,
-		files: nil, // all files are overwritten next time even when timestamp is older
-	}
-}
-
-func (cmd *buildCmd) hasChangedStaticRepos(repos *lockjson.Repos, buildRepos *biRepos, optDir string) bool {
-	if buildRepos == nil { // Full build
-		return true
-	}
-
-	src := pathutil.FullReposPathOf(repos.Path)
-
-	// Get latest mtime of src
-	// TODO: Don't check mtime here, do it when copy altogether
-	srcModTime, err := cmd.getLatestModTime(src)
-	if err != nil {
-		// failed to readdir, do copy again
-		return true
-	}
-
-	if buildRepos.Version == "" {
-		// not found mtime, do copy again
-		return true
-	}
-
-	// Get latest mtime of dst from build-info.json
-	dstModTime, err := time.Parse(time.RFC3339, buildRepos.Version)
-	if err != nil {
-		// failed to parse datetime, do copy again
-		return true
-	}
-
-	return dstModTime.Before(srcModTime)
-}
-
-// Remove ~/.vim/volt/opt/{repos} and copy from ~/volt/repos/{repos}
-func (cmd *buildCmd) updateStaticRepos(repos *lockjson.Repos, done chan actionReposResult) {
-	src := pathutil.FullReposPathOf(repos.Path)
-	dst := pathutil.PackReposPathOf(repos.Path)
-
-	// Remove ~/.vim/volt/opt/{repos}
-	// TODO: Do not remove here, copy newer files only after
-	err := os.RemoveAll(dst)
-	if err != nil {
-		done <- actionReposResult{
-			err:   errors.New("failed to remove repository: " + err.Error()),
-			repos: repos,
-		}
-		return
-	}
-
-	// Copy ~/volt/repos/{repos} to ~/.vim/volt/opt/{repos}
-	buf := make([]byte, 32*1024)
-	si, err := os.Stat(src)
-	if err != nil {
-		done <- actionReposResult{
-			err:   errors.New("failed to copy static directory: " + err.Error()),
-			repos: repos,
-		}
-		return
-	}
-	if !si.IsDir() {
-		done <- actionReposResult{
-			err:   errors.New("failed to copy static directory: source is not a directory"),
-			repos: repos,
-		}
-		return
-	}
-	err = fileutil.TryLinkDir(src, dst, buf, si.Mode(), BuildModeInvalidType)
-	if err != nil {
-		done <- actionReposResult{
-			err:   errors.New("failed to copy static directory: " + err.Error()),
-			repos: repos,
-		}
-		return
-	}
-
-	// Do ":helptags" to generate tags file
-	err = cmd.helptags(repos.Path)
-	if err != nil {
-		done <- actionReposResult{
-			err:   err,
-			repos: repos,
-		}
-		return
-	}
-
-	done <- actionReposResult{
-		err:   nil,
-		repos: repos,
-	}
-}
-
-func (cmd *buildCmd) helptags(reposPath string) error {
-	// Do nothing if <reposPath>/doc directory doesn't exist
-	docdir := filepath.Join(pathutil.PackReposPathOf(reposPath), "doc")
-	if !pathutil.Exists(docdir) {
-		return nil
-	}
-	// Do not invoke vim if not installed
-	_, err := exec.LookPath("vim")
-	if err != nil {
-		return errors.New("vim command is not in PATH: " + err.Error())
-	}
-
-	// Find vim executable from PATH
-	vimExe, err := pathutil.VimExecutable()
-	if err != nil {
-		return err
-	}
-	vimArgs := cmd.makeVimArgs(reposPath)
-
-	// Execute ":helptags doc" in reposPath
-	logger.Debugf("Executing '%s %s' ...", vimExe, strings.Join(vimArgs, " "))
-	err = exec.Command(vimExe, vimArgs...).Run()
-	if err != nil {
-		return errors.New("failed to make tags file: " + err.Error())
-	}
-	return nil
-}
-
-func (*buildCmd) makeVimArgs(reposPath string) []string {
-	return []string{
-		"-u", "NONE", "-N",
-		"-c", "cd " + pathutil.PackReposPathOf(reposPath),
-		"-c", "set rtp+=" + pathutil.PackReposPathOf(reposPath),
-		"-c", "helptags doc", "-c", "quit",
-	}
+	return builder.Build(buildInfo, buildReposMap)
 }
